style: 공백 및 주석 제거
