+++ src/computer/computer.js
@@ -3,7 +3,6 @@ import { NUMBER_LENGTH } from '../libs/constant.js';
 function addRandomNum(answer) {
   // eslint-disable-next-line no-undef
   const randomNum = String(MissionUtils.Random.pickNumberInRange(1, 9));
-
   if (answer.length === 0 || !answer.includes(randomNum)) {
     answer.push(randomNum);
   }
